Fix the GM ".siege" command to use correct castle IDs.
The castles are assigned IDs (castle_id) in the database from 1 through
to 8, and these are the values shown in the help message for the
command. However, internally the code uses zero-based values of 0
through to 7. You need to subtract 1 from the castle ID when using the
siege command if you want the correct castle.