Fix super-past hp/mp regen
The thread was scheduled to run over and over at a specific interval, but when it was set to stop, it just set a flag to false, so it was creating multiple threads all running at the same time.

I changed the calls to actually call the stopHp/MpRegeneration calls that will both set the flag and cancel the timer. It will then re-start when mp/hp needs to regen.