Added birthday functionality, LogStatusUp Fix
The 3.63 client has a birthday of when your character is created. This adds the column to the characters table and ensures it displays correctly directly after creating a character, and any other time the character is loaded.

I also fixed an error in the LogStatusUp class that hadn't had the table name updated.