Added mob count and mob count reset
The new client counts the number of mobs you kill while logged in. It also handles a new SendLocation packet that is used for showing dungeon timers, dragon raid info, and resetting the mob counter.

I've added comments to the SendLocation packet for raids and timers, but they will be implemented at another date