Fix bookmarks doing a random teleport
With the new way bookmarks are loaded, the bookmark call no longer recieves the bookmarkId.

Instead it receives the X, Y, and mapId information directly. We now do a check to verify that bookmark exists before passing on the bookmarkId for existing code to work.

Should probably update the entire code to accept the x, y, and mapid, but this was a quick fix.