Fix bookmark sort and add quicklist
There was an error with the sort query but now bookmarks are alphabetical by default. The client sort has not been coded yet so it is always alphabetical.

The bookmark quicklist is a list at the top of the bookmarks window where users can drag their bookmarks for quick action.