Added delayed packet
With the networking changes, all packets are sent to the client at the same time.

So this causes issues with certain animations forcing timers on their skills. This adds a packet that gets sent on another thread to bypass the limitation.

Dirty hack, but not sure what else to do.